Issue #14909: Renamed section 4.8.4.3 of google style guide implementation
