+++ config/jsoref-spellchecker/whitelist.words
@@ -284,8 +284,8 @@ Ddry
 declarationorder
 Deconstruction
 defau
-defaultcasepresent
 defaultcomeslast
+defaultlabelpresence
 delims
 DENORMALIZER
 Depedency
